% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{AdjRandIndex}
\alias{AdjRandIndex}
\title{Adjusted Rand Index}
\usage{
AdjRandIndex(x, y)
}
\arguments{
\item{x}{a numeric or factor vector or image}

\item{y}{a numeric or factor vector or image}
}
\value{
The Rand index, the Adjusted Rand Index, and the PSNR
}
\description{
Computes the adjusted Rand index and the unadjusted Rand index for two
inputs. These inputs should be binary or categorical and of the same length.
It also computes the PSNR, which is generalized here as simply
-10 log10(Rand).
}
\examples{

x <- rep(1:5, 5)
y <- c(rep(1:5, 4),rep(1,5))
AdjRandIndex(x, y)
}
