% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{binssim}
\alias{binssim}
\title{Categorical Structural Similarity Index Metric (whole image)}
\usage{
binssim(x, y, alpha = 1, beta = 1, gamma = 1, ...)
}
\arguments{
\item{x}{binary or categorical image}

\item{y}{binary or categorical image}

\item{alpha}{normalizing parameter, by default 1}

\item{beta}{normalizing parameter, by default 1}

\item{gamma}{normalizing parameter, by default 1}

\item{...}{Constants can be passed to the components of the index.}
}
\value{
Structural similarity index.
}
\description{
This computes the categorical or binary structural similarity index metric
on a whole-image scale.
}
\examples{
set.seed(20181207)
x <- matrix(sample(1:4, 10000, replace = TRUE), nrow=100)
y <- x
for (i in 1:100) y[i, i] = 1
for (i in 1:99) y[i, i+1] = 1
binssim(x,y)
}
