% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{catmssim_2d}
\alias{catmssim_2d}
\title{Multiscale Categorical Structural Similarity Index Measure (2D)}
\usage{
catmssim_2d(x, y, weights = c(0.0448, 0.2856, 0.3001, 0.2363, 0.1333),
  ...)
}
\arguments{
\item{x}{a binary or categorical image}

\item{y}{a binary or categorical image}

\item{weights}{a vector of weights for the different scales. By default,
five different scales are used.}

\item{...}{additional constants can be passed to internal functions.}
}
\value{
a value less than 1 indicating the similarity between the images.
}
\description{
The categorical structural similary index measure for 2D categorical or binary
images for multiple scales. The default is to compute over 5 scales.
}
\examples{
set.seed(20181207)
x <- matrix(sample(1:4, 128^2, replace = TRUE), nrow=128)
y <- x
for (i in 1:128) y[i, i] = 1
for (i in 1:127) y[i, i+1] = 1
catmssim_2d(x,y)

}
