% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{catssim_2d}
\alias{catssim_2d}
\title{Categorical Structural Similarity Index Measure (2D)}
\usage{
catssim_2d(x, y, ...)
}
\arguments{
\item{x}{a binary or categorical image}

\item{y}{a binary or categorical image}

\item{...}{additional constants can be passed to internal functions.}
}
\value{
a value less than 1 indicating the similarity between the images.
}
\description{
The categorical structural similary index measure for 2D categorical or binary
images for a single scale. This computes it using moving 8x8 windows and is
suitable for modestly-sized images which are not large enough to warrant
looking at multiple scales.
}
\examples{
\dontrun{
set.seed(20181207)
x <- matrix(sample(1:4, 400, replace = TRUE), nrow=20)
y <- x
for (i in 1:20) y[i, i] = 1
for (i in 1:19) y[i, i+1] = 1
catssim_2d(x,y)
}
}
\keyword{internal}
