% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{catssim_2d_cpp}
\alias{catssim_2d_cpp}
\title{Categorical Structural Similarity Index Measure (2D) - C++ version}
\usage{
catssim_2d_cpp(x, y, window = c(11, 11), method = "Cohen", ...)
}
\arguments{
\item{x, y}{binary or categorical image}

\item{window}{the window size over which the similarity is computed.
The default is c(11, 11).}

\item{method}{whether to use Cohen's kappa (\code{Cohen}),
Jaccard Index (\code{Jaccard}), Dice index (\code{Dice}),
accuracy (\code{accuracy}),  Rand index (\code{Rand}),
Adjusted Rand Index (\code{AdjRand} or \code{ARI}), normalized mutual
information (\code{NMI} or \code{MI}) or the adjusted mutual
information, \code{AMI} and \code{ami}, as
the similarity index. Note Jaccard and Dice should only be used on
binary data.}

\item{...}{additional constants can be passed to internal functions.}
}
\value{
a value less than 1 indicating the similarity between the images.
}
\description{
This is a C++ implementation of catssim_2d for improved performance.
Computes the categorical or binary structural similarity index metric
on a sliding window basis.
}
\examples{
set.seed(20181207)
x <- matrix(sample(1:4, 400, replace = TRUE), nrow = 20)
y <- x
for (i in 1:20) y[i, i] <- 1
for (i in 1:19) y[i, i + 1] <- 1
catssim_2d_cpp(x, y)
}
