% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{gini}
\alias{gini}
\title{Gini-Simpson index}
\usage{
gini(x)
}
\arguments{
\item{x}{binary or categorical image or vector}
}
\value{
Gini-Simpson index (between 0 and 1)
}
\description{
A measure of diversity that goes by a number of different names, such as
the probability of interspecific encounter or the Gibbs-Martin index.
It is \eqn{1 - sum(p_i^2)}, where \eqn{p_i} is the probability of observing class i.
}
\examples{

x <- rep(c(1:4),5)
gini(x)


}
