% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{ginicorr}
\alias{ginicorr}
\title{Corrected Gini-Simpson index}
\usage{
ginicorr(x, k)
}
\arguments{
\item{x}{binary or categorical image or vector}

\item{k}{number of categories}
}
\value{
Gini-Simpson index (corrected based on the number of categories
so that max possible is 1)
}
\description{
Corrected Gini-Simpson index
}
\examples{

x <- rep(c(1:4),5)
ginicorr(x, 4)

}
