% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{hamming}
\alias{hamming}
\title{Accuracy/Hamming Index}
\usage{
hamming(x, y)
}
\arguments{
\item{x, y}{binary image or vector}
}
\value{
Accuracy (Hamming index). Proportion of pixels (voxels) that agree.
}
\description{
Accuracy/Hamming Index
}
\examples{
\dontrun{
x <- rep(c(0,1), 7)
y <- rep(c(1,1), 7)
hamming(x,y)
}
}
\keyword{internal}
