% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard Index}
\usage{
jaccard(x, y)
}
\arguments{
\item{x}{binary image or vector}

\item{y}{binary image or vector}
}
\value{
Jaccard index - note this privileges 1 over 0
}
\description{
Jaccard Index
}
\examples{
\dontrun{
x <- rep(c(0,1), 7)
y <- rep(c(1,1), 7)
jaccard(x,y)
}
}
\keyword{internal}
