% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{methodparser}
\alias{methodparser}
\title{Method Parser}
\usage{
methodparser(method)
}
\arguments{
\item{method}{The method used as a similarity metric. Certain abbreviations work.
\code{Cohen}, \code{cohen}, \code{C}, \code{c}, \code{Kappa} and \code{kappa} yield Cohen's kappa.
\code{AdjRand}, \code{adjrand}, \code{Adj}, \code{adj}, \code{a}, \code{A}, \code{ARI}, and \code{ari} yield the adjusted Rand index.
\code{Rand}, \code{rand}, \code{r}, and \code{R} yield the Rand index.
\code{Jaccard}, \code{jaccard}, \code{j}, and \code{J} yield the Jaccard index.
\code{Dice}, \code{dice}, \code{D}, and \code{d} yield the Dice index.
\code{Accuracy}, \code{accuracy}, \code{Hamming}, \code{hamming}, \code{H}, and \code{h} yield the accuracy.}
}
\value{
the name of the similarity metric.
}
\description{
Parses input of method to a standardized name
}
\keyword{internal}
