% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{sqrtgini}
\alias{sqrtgini}
\title{Modified Gini-Simpson index}
\usage{
sqrtgini(x)
}
\arguments{
\item{x}{binary or categorical image or vector}
}
\value{
Modified Gini-Simpson index (square root of squared frequencies)
}
\description{
The Gini-Simpson index, except with the square root of the summed squared
probabilities.
}
\examples{

x <- rep(c(1:4),5)
sqrtgini(x)

}
