% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{sqrtginicorr}
\alias{sqrtginicorr}
\title{Modified Corrected Gini index}
\usage{
sqrtginicorr(x, k)
}
\arguments{
\item{x}{binary or categorical image or vector}

\item{k}{number of categories}
}
\value{
Modified corrected Gini index (square root of squared frequencies) -
max possible is 1.
}
\description{
The Gini-Simpson index, except with the square root of the summed squared
probabilities.
}
\examples{

x <- rep(c(1:4),5)
sqrtginicorr(x, 4)

}
