% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{AdjRandIndex}
\alias{AdjRandIndex}
\title{Adjusted Rand Index}
\usage{
AdjRandIndex(x, y)
}
\arguments{
\item{x}{a numeric or factor vector or image}

\item{y}{a numeric or factor vector or image}
}
\value{
The Rand index and the Adjusted Rand Index
}
\description{
Computes the adjusted Rand index and the unadjusted Rand index for two
inputs.
}
\examples{

x <- rep(1:5, 5)
y <- c(rep(1:5, 4),rep(1,5))
AdjRandIndex(x, y)
}
