% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{AdjRandIndex}
\alias{AdjRandIndex}
\title{Adjusted Rand Index}
\usage{
AdjRandIndex(x, y)
}
\arguments{
\item{x}{a numeric or factor vector or image}

\item{y}{a numeric or factor vector or image}
}
\value{
The Jaccard index, the Adjusted Rand Index, the PSNR, and Cohen's Kappa. Note:
The Jaccard index will not make sense if this is not binary.
}
\description{
Computes the adjusted Rand index for two
inputs. These inputs should be binary or categorical and of the same length.
It also computes the PSNR, which is generalized here as simply
\eqn{-10 log_{10}(MSE)}. The adjusted Rand index, Jaccard Index, and Cohen's Kappa are used as a measure of
the similarity of the structure of the two images. A small constant is added to the numerator
and denominator of the Adjusted Rand index to ensure stability, as it is possible to have a zero
denominator. The PSNR can be infinite if the error rate is 0.
}
\examples{

x <- rep(1:5, 5)
y <- c(rep(1:5, 4),rep(1,5))
AdjRandIndex(x, y)
}
\references{
Lawrence Hubert and Phipps Arabie (1985).
"Comparing partitions". Journal of Classification. 2 (1): 193–218. \doi{10.1007/BF01908075}

W. M. Rand (1971). "Objective criteria for the evaluation of clustering methods".
Journal of the American Statistical Association. American Statistical Association. 66 (336): 846–850.
\doi{10.2307/2284239}

Cohen, Jacob (1960). "A coefficient of agreement for nominal scales".
Educational and Psychological Measurement. 20 (1): 37–46. \doi{10.1177/001316446002000104}
}
