% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{AdjRandIndex}
\alias{AdjRandIndex}
\title{Adjusted Rand Index and other similarity indexes (Deprecated)}
\usage{
AdjRandIndex(x, y)
}
\arguments{
\item{x, y}{a numeric or factor vector or image}
}
\value{
The accuracy, Jaccard index, the Adjusted Rand Index,
the Rand index, the PSNR, Cohen's Kappa,
normalized mutual information (NMI) and adjusted mutual information (AMI).
Note: The Jaccard index will not make sense if this is not binary.
}
\description{
Computes the adjusted Rand index and several other similarity measures for
two inputs. These inputs should be binary or categorical and of the same
length.
It also computes the PSNR, which is generalized here as simply
\eqn{-10 log_{10}(MSE)}.
The adjusted Rand index, Jaccard Index, Cohen's Kappa,
normalized mutual information (NMI) and adjusted mutual information (AMI)
are used as a measure of the similarity of the structure of the two images.
A small constant is added to the numerator
and denominator of the Adjusted Rand index to ensure stability,
as it is possible to have a zero
denominator. The normalized mutual information is defined here as:
\eqn{2H(X,Y)/(H(X)+H(Y)),} but is set to be 0 if both H(X) and H(Y) are 0.
The PSNR can be infinite if the error rate is 0.
The Jaccard index and accuracy can
account for NA values, but the Adjusted Rand and Cohen's Kappa cannot.
}
\references{
W. M. Rand (1971).
"Objective criteria for the evaluation of clustering methods".
Journal of the American Statistical Association.
American Statistical Association. 66 (336): 846–850.
\doi{10.2307/2284239}

Lawrence Hubert and Phipps Arabie (1985).
"Comparing partitions". Journal of Classification. 2 (1): 193–218.
\doi{10.1007/BF01908075}

Cohen, Jacob (1960). "A coefficient of agreement for nominal scales".
Educational and Psychological Measurement. 20 (1): 37–46.
\doi{10.1177/001316446002000104}

Jaccard, Paul (1912). "The distribution of the flora in the alpine zone,”
New Phytologist, vol. 11, no. 2, pp. 37–50.
\doi{10.1111/j.1469-8137.1912.tb05611.x}

Nguyen Xuan Vinh, Julien Epps, and James Bailey (2010).
Information Theoretic Measures for Clusterings Comparison:
Variants, Properties, Normalization and Correction for Chance.
J. Mach. Learn. Res. 11 (December 2010), 2837–2854.
\doi{10.5555/1756006.1953024}
}
