% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{catmssim_3d_cube}
\alias{catmssim_3d_cube}
\title{Multiscale Categorical Structural Similarity Index Measure for a Cube (3D)}
\usage{
catmssim_3d_cube(x, y, weights = c(0.0448, 0.2856, 0.3001, 0.2363,
  0.1333), ...)
}
\arguments{
\item{x}{a binary or categorical image}

\item{y}{a binary or categorical image}

\item{weights}{a vector of weights for the different scales. By default,
five different scales are used.}

\item{...}{additional constants can be passed to internal functions.}
}
\value{
a value less than 1 indicating the similarity between the images.
}
\description{
The categorical structural similary index measure for 3D categorical or binary
images for multiple scales. The default is to compute over 5 scales.
This computes a 3D measure
}
\examples{
x <- array(sample(1:4, 16^3, replace = TRUE), dim = c(16,16,16))
y <- x
for (j in 1:16){
for (i in 1:16) y[i, i, j] = 1
for (i in 1:15) y[i, i+1, j] = 1
}
catmssim_3d_cube(x,y, weights = c(.75,.25))
}
