% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{catmssim_3d_cube}
\alias{catmssim_3d_cube}
\title{Multiscale Categorical Structural Similarity Index Measure for a Cube (3D)}
\usage{
catmssim_3d_cube(x, y, weights = c(0.0448, 0.2856, 0.3001, 0.2363,
  0.1333), window = 4, ...)
}
\arguments{
\item{x}{a binary or categorical image}

\item{y}{a binary or categorical image}

\item{weights}{a vector of weights for the different scales. By default,
five different scales are used.}

\item{window}{size of window, by default 4}

\item{...}{additional constants can be passed to internal functions.}
}
\value{
a value less than 1 indicating the similarity between the images.
}
\description{
The categorical structural similary index measure for 3D categorical or binary
images for multiple scales. The default is to compute over 5 scales.
This computes a 3D measure based on 4x4x4 windows by default with 5 levels of downsampling.
}
\examples{
set.seed(20181207)
dim = 16
x <- array(sample(1:4, dim^3, replace = TRUE), dim = c(dim,dim,dim))
y <- x
for (j in 1:dim){
for (i in 1:dim) y[i, i, j] = 1
for (i in 1:(dim-1)) y[i, i+1, j] = 1
}
catmssim_3d_cube(x,y, weights = c(.75,.25))
}
