% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim.R
\name{catmssim_3d_slice}
\alias{catmssim_3d_slice}
\title{Multiscale Categorical Structural Similarity Index Measure by Slice (3D)}
\usage{
catmssim_3d_slice(x, y, weights = c(0.0448, 0.2856, 0.3001, 0.2363,
  0.1333), window = 8, method = "Cohen", ...)
}
\arguments{
\item{x}{a binary or categorical image}

\item{y}{a binary or categorical image}

\item{weights}{a vector of weights for the different scales. By default,
five different scales are used.}

\item{window}{window size, by default 8.}

\item{method}{whether to use Cohen's kappa, Jaccard Index, or Adjusted Rand Index as
the similarity index. Note Jaccard should only be used on binary data.}

\item{...}{additional constants can be passed to internal functions.}
}
\value{
a value less than 1 indicating the similarity between the images.
}
\description{
The categorical structural similary index measure for 3D categorical or binary
images for multiple scales. The default is to compute over 5 scales.
This computes a 2D measure for each x-y slice of the z-axis
and then averages over the z-axis.
}
\examples{
set.seed(20181207)
dim = 16
x <- array(sample(1:4, dim^3, replace = TRUE), dim = c(dim,dim,dim))
y <- x
for (j in 1:dim){
for (i in 1:dim) y[i, i, j] = 1
for (i in 1:(dim-1)) y[i, i+1, j] = 1
}
catmssim_3d_slice(x,y, weights = c(.75,.25))
}
