% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim-package.R
\docType{package}
\name{catsim-package}
\alias{catsim}
\alias{catsim-package}
\title{catsim: Binary and Categorical Image Structural Similarity Metric}
\description{
Computes a structural similarity metric (after the style of MS-SSIM for images)
    for binary and categorical 2D and 3D images.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/gzt/catsim}
  \item \url{https://gzt.github.io/catsim}
  \item Report bugs at \url{https://github.com/gzt/catsim/issues}
}

}
\author{
\strong{Maintainer}: Geoffrey Thompson \email{gzthompson@gmail.com} (0000-0003-2436-8822)

}
\keyword{internal}
