% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catsim-package.R
\docType{package}
\name{catsim-package}
\alias{catsim-package}
\alias{_PACKAGE}
\title{CatSIM: Binary and Categorical Image Structural Similarity Metric}
\description{
Computes a structural similarity metric (after the style of MS-SSIM for images)
for binary and categorical 2D and 3D images. Can be based on accuracy (simple matching),
Cohen's kappa, Rand index, adjusted Rand index, Jaccard index, Dice index, normalized
mutual information, or adjusted mutual information. In addition, has fast computation
of Cohen's kappa, the Rand indices, and the two mutual informations.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/gzt/catsim}
  \item \url{https://gzt.github.io/catsim}
  \item Report bugs at \url{https://github.com/gzt/catsim/issues}
}

}
\author{
\strong{Maintainer}: Geoffrey Thompson \email{gzthompson@gmail.com} (\href{https://orcid.org/0000-0003-2436-8822}{ORCID})

}
\keyword{internal}
